+++ src/lib/events/eventActorSetMovementSpeed.js
@@ -28,6 +28,75 @@ const fields = [
     type: "moveSpeed",
     defaultValue: 1,
   },
+  {
+    type: "group",
+    wrapItems: true,
+    conditions: [
+      {
+        sceneType: ["platform", "adventure"],
+      },
+      {
+        or: [
+          [
+            {
+              key: "actorId",
+              eq: "$self$",
+            },
+            {
+              entityTypeNot: ["actor"],
+            },
+          ],
+          [
+            {
+              key: "actorId",
+              eq: "player",
+            },
+          ],
+        ],
+      },
+    ],
+    fields: [
+      {
+        label: l10n("FIELD_SET_PLAYER_MOVEMENT_PLATFORMER_ADVENTURE_WARNING"),
+        labelVariant: "warning",
+        flexBasis: "100%",
+      },
+      {
+        type: "addEventButton",
+        hideLabel: true,
+        label: l10n("EVENT_ENGINE_FIELD_SET"),
+        defaultValue: {
+          id: "EVENT_ENGINE_FIELD_SET",
+          values: {
+            engineFieldKey: "plat_walk_vel",
+          },
+          replace: true,
+        },
+        conditions: [
+          {
+            sceneType: ["platform"],
+          },
+        ],
+      },
+      {
+        type: "addEventButton",
+        hideLabel: true,
+        label: l10n("EVENT_ENGINE_FIELD_SET"),
+        defaultValue: {
+          id: "EVENT_ENGINE_FIELD_SET",
+          values: {
+            engineFieldKey: "adv_walk_vel",
+          },
+          replace: true,
+        },
+        conditions: [
+          {
+            sceneType: ["adventure"],
+          },
+        ],
+      },
+    ],
+  },
 ];
 
 const compile = (input, helpers) => {