+++ src/lib/events/eventTextDialogue.js
@@ -27,6 +27,17 @@ const fields = [
     },
   },
 
+  {
+    label: l10n("FIELD_TEXT_IN_LOGO_WARNING"),
+    labelVariant: "warning",
+    flexBasis: "100%",
+    conditions: [
+      {
+        sceneType: ["logo"],
+      },
+    ],
+  },
+
   // Text Section
 
   {